+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added OPD file format (OPD files may have differing values of step_x and step_y, this could cause unexpected errors 
   in the calculation of some parameters)
 - Added XYZ file format. The reader assumes units of meters and data points on a regular grid.
+- Added methods to level and calculate stepheight and cavity volume for rectangular ablation craters
 ## v0.5.1
 - Changed plot parameter of Surface.depth to plot a specific or multiple profiles
 - Fixed hashing of mutable types for caching of method calls with mutable parameters